% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAgeLen.R
\name{calcAgeLen}
\alias{calcAgeLen}
\title{calcAgeLen}
\usage{
calcAgeLen(requested = NULL)
}
\arguments{
\item{requested}{The default value is \code{NULL}.}
}
\description{
This function calculates the ages and lengths.
}
\seealso{
Other Gale-force: \code{\link{calcNumsWeights}},
  \code{\link{extractData}}, \code{\link{getUserInput}},
  \code{\link{stranal}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
