% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\title{extractData}
\usage{
extractData(requested = NULL, agency = NULL, dfSpp = NULL, type = NULL,
  missions = NULL, strata = NULL, bySex = NULL)
}
\arguments{
\item{requested}{The default value is \code{NULL}. This determines which 
analytic will be performed.}

\item{agency}{The default value is \code{NULL}.}

\item{dfSpp}{The default value is \code{NULL}.}

\item{type}{The default value is \code{NULL}.}

\item{missions}{The default value is \code{NULL}.}

\item{strata}{The default value is \code{NULL}.}

\item{bySex}{The default value is \code{NULL}.}
}
\description{
This function doesn't do anything yet.
}
\seealso{
Other Gale-force: \code{\link{calcAgeLen}},
  \code{\link{calcNumsWeights}},
  \code{\link{getUserInput}}, \code{\link{stranal}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
