% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserInput.R
\name{getUserInput}
\alias{getUserInput}
\title{getUserInput}
\usage{
getUserInput(requested = NULL)
}
\arguments{
\item{requested}{The default value is \code{NULL}.}
}
\description{
This function can present the user with a variety of select 
boxes from which they can select appropriate options.  The select box that is
shown is determined by the value of the paramter \code{"requested"}.
Some of the options that are presented are dependent on the existence of 
values written to the global environment by the stranal function.
}
\seealso{
Other Gale-force: \code{\link{calcAgeLen}},
  \code{\link{calcNumsWeights}}, \code{\link{extractData}},
  \code{\link{stranal}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
