% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stranal.R
\name{stranal}
\alias{stranal}
\title{stranal}
\usage{
stranal(usepkg = "roracle")
}
\arguments{
\item{usepkg}{The default value is \code{'roracle'}, but \code{'rodbc'} is works as well.
This describes the R package you use to connect to Oracle.}
}
\description{
Stuff.
}
\seealso{
Other Gale-force: \code{\link{calcAgeLen}},
  \code{\link{calcNumsWeights}}, \code{\link{extractData}},
  \code{\link{getUserInput}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}l
}
