% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratisfy.R
\name{stratisfy}
\alias{stratisfy}
\title{stratisfy}
\usage{
stratisfy(
  usepkg = "rodbc",
  agency = "DFO",
  type = 1,
  year = NULL,
  season = NULL,
  missions = NULL,
  strataTable = NULL,
  wingspread = 41,
  towDist = 1.75,
  strata = NULL,
  areas = NULL,
  spp = NULL,
  bySex = NULL,
  ageBySex = FALSE,
  output = "new",
  alkTable = NULL,
  file_id = NULL,
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_"
)
}
\arguments{
\item{usepkg}{The default value is \code{'rodbc'}, but \code{'roracle'} 
works as well. This describes the R package you use to connect to Oracle.}

\item{agency}{The default value is \code{'DFO'}, the other option is 
\code{'NMFS'}.  Setting to \code{NULL} will result in a pick list.}

\item{type}{This is the 'experiment type', and the default value is \code{1}.
For \code{agency='DFO'}, \code{5} is also acceptable. For agency = NMFS, 136 
is probable, but you can type any integer, and the typed number will be 
embedded in\code{USNEFSC.USS_STATION.SHG <= type}. Setting to \code{NULL} 
will result in a pick list.}

\item{year}{The default value is \code{NULL}, which will result in a pick list.
An example of a valid, non-empty value is \code{2017}.}

\item{season}{The default value is \code{''}, which will result in a pick 
list. Valid values are \code{'SUMMER', 'SPRING', 'WINTER', and 'FALL'}, 
subject to the availability of data in that season for the selected agency, 
year, type, etc...}

\item{missions}{The default value is \code{NULL}, which will result in a pick 
list. An example valid value is \code{c('NED2016016')}.  If your choice is 
not valid (given your other selections), you will be presented with a pick
list of valid options.}

\item{strataTable}{The default value is \code{NULL}.
Depending on your data, other valid values may include: 
\code{'GROUNDFISH.GSSTRATUM', 'USNEFSC.DFO5ZJM','USNEFSC.DFO5ZGHNO','USNEFSC.NMFS5ZJM',
'USNEFSC.NMFS5ZGHNO','USNEFSC.NMFS5ZJMC','USNEFSC.NMFS5ZJMU',
'USNEFSC.NMFS5ZU'}.}

\item{wingspread}{This is the width in feet of a standard tow.  The default 
value is \code{41}.  Setting this to \code{NULL} will result in a pick list.}

\item{towDist}{This is the length (NM) of a standard tow.  The default value 
is \code{1.75}.  Setting this to \code{NULL} will result in a pick list.}

\item{strata}{These are the strata for which you want results.  The default 
value is \code{NULL} The DFO SUmmer survey are c(440:495).}

\item{areas}{These are the areas for which you want results.  Some strata 
can be further broken down by areas, so this allows selection of data within 
a part of a strata. It can also be set to \code{'all'} or a vector of valid areas 
(e.g. \code{c('424','511','4xs')})}

\item{spp}{This is the species code for the species you want to analyze.
The default value is \code{NULL}, which will result in a pick list. An example 
of a valid, non-empty value is \code{2526}.}

\item{bySex}{If the selected species was measured by sex, this parameter
allows you to perform calculations by sex.  The default value is \code{NULL}, 
which will result in a pick list. An example of a valid, non-empty value is 
\code{TRUE}.}

\item{ageBySex}{The APL version of stratisfy ignored sex differences in some 
age results (e.g. 'age by set', 'age mean', 'age total', ' etc.) despite
analyses being done by sex.  The default value of this parameter is 
\code{FALSE} so that the original results are emulated.  However, setting 
this to \code{TRUE} (when \code{bySex=TRUE}), will show age results broken 
down by sex.  Setting this to \code{''} will result in a pick list.}

\item{output}{The default value is \code{'new'}.  This determines the format 
of the output Excel file. Setting this to \code{'classic'} will emulate the
original APL stratisfy results, including overriding your parameter for 
\code{ageBySex} and forcing it to FALSE.  If no excel output is desired, set
this parameter to an empty string \code{''}}

\item{alkTable}{The default is \code{NULL}.  Setting this to a valid path
allows you to add additional values for the age length key.}

\item{file_id}{This is a modifier you can have appended to the output 
filename.  The resultant file will be called \code{'Mar_stratisfy<_file_id>.xlsx'}}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for this stored
in your environment (e.g. from an rprofile file), this can be left and that
value will be used.  If a value for this is provided, it will take priority
over your existing value.}
}
\description{
stratisfy is a program capable of Stratified Analyses of both DFO 
and NMFS fisheries survey data. This package attempts to emulate the results
of the older, APL version which has been in use for many years.

Using stratisfy requires: 
\enumerate{
\item an Oracle account on the DFO PTRAN database;
\item access to PTRAN (i.e. either within the network, or via a VPN) 
\item appropriate select permissions on the 'groundfish' and/or 'usnefsc' 
schemas
}
New Oracle accounts are requested by filling out a computer account request 
form. These forms may be obtained from IM&TS. Select access to the DFO and 
NMFS data is provided by the respective datasets manager (checking with 
\email{Mike.McMahon@dfo-mpo.gc.ca} is a good start).

Users can supply as much or as little information as they want when calling 
stratisfy(). If required information is not supplied, a select box will show 
the available options to the user. Please see below for examples of function 
calls providing both the minimum and the maximum number of parameters.

Fun Facts About stratisfy
\enumerate{
\item Berried Females -- APL stratisfy was inconsistent with species where 
females could be coded as 'berried' (i.e. 3).  When unsexed analyses were 
done, the females were included, but when sexed analyses were done, these 
females were not included in the results.  Such inconsistency seems 
suboptimal, and is retained for now such that APL and R versions of the 
application can be compared. 
\item Ages By Sex -- When APL stratisfy did analyses by sex, it combined the 
sexes for the 'Age Mean', 'Age Total', (and standard errors of each).  This 
version adds the parameter \code{ageBySex} so that by setting it to FALSE, 
you can get the combined results, but you can also set it to TRUE, and get 
the values for each sex individually.
\item I liked the old way -- Some aspects of the 'classic' (i.e. APL) 
stratisfy seemed inefficient.   For example, the Strata information was broken 
up over multiple worksheets, as was the weight information. By default, a new 
spreadsheet has been designed, but since there will always be those who 
prefer the old way, setting \code{output='classic'} will generate a 
spreadsheet almost indistinguishable from the APL version.
}
}
\note{
Reports of errors and/or recommendations for improvements should be 
addressed to:
Mike McMahon (Mike.McMahon@dfo-mpo.gc.ca)
Fisheries and Oceans Canada
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{Gale-force}
