% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNumsWeights.R
\name{calcNumsWeights}
\alias{calcNumsWeights}
\title{calcNumsWeights}
\usage{
calcNumsWeights(requested = NULL)
}
\arguments{
\item{requested}{The default value is \code{NULL}.}
}
\description{
This function calculates the numbers and weights.
}
\seealso{
Other Gale-force: \code{\link{calcAgeLen}},
  \code{\link{extractData}}, \code{\link{getUserInput}},
  \code{\link{stranal}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
