% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\title{extractData}
\usage{
extractData(requested = NULL)
}
\arguments{
\item{requested}{The default value is \code{NULL}.}
}
\description{
This function doesn't do anything yet.
}
\seealso{
Other Gale-force: \code{\link{calcAgeLen}},
  \code{\link{calcNumsWeights}},
  \code{\link{getUserInput}}, \code{\link{stranal}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
