% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserInput.R
\name{getUserInput}
\alias{getUserInput}
\title{getUserInput}
\usage{
getUserInput(requested = NULL, agency = NULL, type = NULL,
  strataTable = NULL, year = NULL, season = NULL, wingspread = NULL,
  towDist = NULL, spp = NULL, bySex = NULL, strata = NULL,
  dfMissionsStrata = NULL, ageBySex = NULL)
}
\arguments{
\item{requested}{The default value is \code{NULL}. This determines which 
picklist will be invoked.}

\item{agency}{The default value is \code{NULL}.  Valid values are 'DFO' and
'NMFS'.}

\item{type}{The default value is \code{NULL}.  Setting to \code{NULL} will 
result in a pick list.}

\item{strataTable}{The default value is \code{NULL}.  Setting to 
\code{NULL} will result in a pick list.}

\item{year}{The default value is \code{NULL}.  Setting to \code{NULL} will 
result in a pick list.}

\item{season}{The default value is \code{NULL}.  Setting to \code{NULL} 
will result in a pick list.}

\item{wingspread}{The default value is \code{NULL}.  Setting to \code{NULL} 
will result in a pick list.}

\item{towDist}{The default value is \code{NULL}.  Setting to \code{NULL} 
will result in a pick list.}

\item{spp}{The default value is \code{NULL}.  Setting to \code{NULL} will 
result in a pick list.}

\item{bySex}{The default value is \code{NULL}.  Setting to \code{NULL} will 
result in a pick list.}

\item{strata}{The default value is \code{NULL}.  Setting to \code{NULL} 
will result in a pick list.}

\item{dfMissionsStrata}{The default value is \code{NULL}.  Setting to 
\code{NULL} will result in a pick list.}

\item{ageBySex}{The default value is \code{NULL}.  Setting to 
\code{NULL} will result in a pick list.}
}
\description{
This function can present the user with a variety of select 
boxes from which they can select appropriate options.  The select box that is
shown is determined by the value of the paramter \code{"requested"}.
Some of the options that are presented are dependent on the existence of 
values written to the global environment by the stranal function.
}
\seealso{
Other Gale-force: \code{\link{calcAgeLen}},
  \code{\link{calcNumsWeights}}, \code{\link{extractData}},
  \code{\link{stranal}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
